price_optimization_backend/requirements.txt
Django==4.2.18 
django-cors-headers==4.4.0  
djangorestframework==3.15.2
django-allauth==65.3.1
django-redis==5.4.0